abstraction of muzero config
